Revert "Dam you machine"
This reverts commit 8c96b85d36bb79ec1794477d5489941e819add5b.

+++ Assets/_src/Scripts/SplineMovements/SplineMovements.Data/SplineMoveComponent.cs
@@ -7,7 +7,7 @@ namespace _src.Scripts.SplineMovements.SplineMovements.Data
     {
         public byte CurveIndex;
         public half Speed;
-        public float Distance;
+        public half Distance;
         public half Lerp;
     }
 }
